### 2025-02-20 #7811 @editorialbot
[PRE REVIEW]: BREADR: An R Package for the Bayesian Estimation of Genetic Relatedness from Low-coverage Genotype Data
**Submitting author:** @BenRohrlach (<a href="http://orcid.org/0000-0002-4204-5018">Adam Rohrlach</a>)
**Repository:** https://github.com/jonotuke/BREADR
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
