+++ BREADR.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={216},
   year={2024},
-  publisher={Springer}
+  publisher={Springer},
+  doi = "10.1186/s13059-024-03350-3"
 }
 
 @article{fernandes:2021,
@@ -17,7 +18,8 @@
   number={1},
   pages={21262},
   year={2021},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London}, 
+  doi = "10.1038/s41598-021-00581-3"
 }
 
 @article{ghalichi:2024,
@@ -26,7 +28,8 @@
   journal={Nature},
   pages={1--9},
   year={2024},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London}, 
+  doi = "10.1038/s41586-024-08113-5"
 }
 
 @article{gretzinger:2024,
@@ -35,7 +38,8 @@
   journal={Nature Human Behaviour},
   pages={1--14},
   year={2024},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi = "10.1038/s41562-024-01888-7"
 }
 
 @article{kennett:2017,
@@ -46,7 +50,8 @@
   number={1},
   pages={14115},
   year={2017},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi = "10.1038/ncomms14115"
 }
 
 @article{lipatov:2015,
@@ -55,7 +60,8 @@
   journal={BioRxiv},
   pages={023374},
   year={2015},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi = "10.1101/023374"
 }
 
 @article{monroy:2018,
@@ -66,7 +72,8 @@
   number={4},
   pages={e0195491},
   year={2018},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA}, 
+  doi = "10.1371/journal.pone.0195491"
 }
 
 @article{penske:2024,
@@ -77,7 +84,8 @@
   number={1},
   pages={3871},
   year={2024},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi = "10.1038/s41598-024-54462-6"
 }
 
 @article{popli:2023,
@@ -88,7 +96,8 @@
   number={1},
   pages={10},
   year={2023},
-  publisher={Springer}
+  publisher={Springer},
+  doi = "10.1186/s13059-023-02847-7"
 }
 @article{pilli:2024,
   title={Ancient DNA challenges prevailing interpretations of the Pompeii plaster casts},
@@ -98,7 +107,8 @@
   number={22},
   pages={5307--5318},
   year={2024},
-  publisher={Elsevier}
+  publisher={Elsevier}, 
+  doi = "10.1016/j.cub.2024.10.007"
 }
 
 @article{queller:1989,
@@ -109,15 +119,20 @@
   number={2},
   pages={258--275},
   year={1989},
-  publisher={Blackwell Publishing Inc Malden, USA}
+  publisher={Blackwell Publishing Inc Malden, USA},
+  doi = "10.1111/j.1558-5646.1989.tb04226.x"
 }
 
 @article{ringbauer:2023,
-  title={ancIBD-Screening for identity by descent segments in human ancient DNA},
-  author={Ringbauer, Harald and Huang, Yilei and Akbari, Ali and Mallick, Swapan and Patterson, Nick and Reich, David},
-  journal={BioRxiv},
-  year={2023},
-  publisher={Cold Spring Harbor Laboratory Preprints}
[…]
