### 2025-03-29 #7916 @nilabhrardas
Hi @BenRohrlach,

I had the opportunity to review the paper and would like to share some thoughts. I found it incredibly well written and the inclusion of the contextualised derivation of the posterior probability was a nice touch.

Regarding the package, the manual is quite well documented and easy to follow for first time users. I was able to reproduce the example results. However, I have a couple of reservations:
1. The example pipeline uses "a real (anonymised) ancient DNA data set" which is available in the github repository. However, I was unable to find any details regarding the dataset itself - whether it is an original dataset/published/publicly available. I would suggest inhcluding further details about the data adhering to the [JOSS Policies](https://joss.readthedocs.io/en/latest/policies.html#joss-policies) if an original dataset, and including a reference or link to the original publication otherwise - similar to how the use of a simulated genotype data has been mentioned in the paper.
2. I suggest incorporating the example pipeline into a standalone vignette that is accessible from within RStudio.

Overall, the paper is well written and the package is well documented. @spholmes, I am happy to recommend this paper for acceptance given the above comments are addressed.

Lastly, I have a few recommendations (not suggesting you must address these) regarding the README documentation. Firstly, using "path/to/eigenstat/indfile", which is a widely used path holder, instead of "path_to_eigenstrat_indfile" which is presently used, and lastly, separating the code and the output into two snippets so that the click-to-copy function does not copy both the code and the (commented out) output lines to the clipboard.

Hope these were helpful.

Best,
@nilabhrardas
