### 2025-03-29 jonotuke/BREADR#3 @xin-huang
[JOSS REVIEW] Suggestions for Documentation
This is part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/7916)

**Installation instructions:** According to the reference manual (https://cran.r-project.org/web/packages/BREADR/BREADR.pdf), the package requires R version ≥ 2.10. However, I was unable to install the package from CRAN using R 4.1.2.

It would be helpful to include the link to the package website (https://jonotuke.github.io/BREADR/) in the README file, so that users can more easily locate the reference manual and other documentation.

**Example usage:** It would be helpful to include a direct link to the dataset for easier testing.

https://github.com/jonotuke/BREADR/blob/36b86a2c925ca92b49e882629fb404e7f85f281d/README.md?plain=1#L72-L75

For example,

```
 Note that since the Eigenstrat files are too large to include, we
 provide a pre-processed data set called [counts_example](https://github.com/jonotuke/BREADR/blob/master/data/counts_example.rda) which is the
 product of running processEigenstrat() on real data (although the raw
 Eigenstrat files can be found on [the Github page](https://github.com/jonotuke/BREADR/tree/master/data-raw)).
```

**Automated tests:** Automated tests are present, but currently failing. I recommend the authors fix the test suite.

**Community guidelines:** There are no clear guidelines for contributing to the software. I suggest adding contribution guidelines to encourage community involvement.
