### 2025-03-29 jonotuke/BREADR#4 @xin-huang
[JOSS REVIEW] Suggestions for Software Paper
This is part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/7916)

In the paper, the authors mention that simulated data were used for example usage; however, it is unclear how these data were generated. I suggest the authors provide documentation on the data simulation process to improve reproducibility.

https://github.com/jonotuke/BREADR/blob/36b86a2c925ca92b49e882629fb404e7f85f281d/paper.md?plain=1#L38

Is the simulated dataset located at https://github.com/jonotuke/BREADR/tree/master/data-raw? If so, it would be helpful to include a link to it in the paper.

Ideally, the authors could also share the full set of scripts used for the analysis presented in the paper, which would make reproduction easier and more convenient.
