### 2025-03-29 #7916 @xin-huang
In this manuscript, the authors present an R package for estimating genetic relatedness from sequencing data, with a particular focus on extremely low-coverage data. This has potential utility in ancient DNA research. I have several suggestions and have opened two corresponding GitHub issues: https://github.com/jonotuke/BREADR/issues/3 and https://github.com/jonotuke/BREADR/issues/4.

Additionally, based on [the contribution history](https://github.com/jonotuke/BREADR/graphs/contributors), it appears that the second author (@jonotuke) is responsible for all of the code. It is therefore unclear whether the submitting author (@BenRohrlach) has made a major contribution to the software.
