### 2025-04-01 bcbfd32
Added data description and contribution to paper
