+++ BREADR.bib
@@ -54,6 +54,19 @@
   doi = "10.1038/ncomms14115"
 }
 
+@article{li2009fast,
+	author = {Li, Heng and Durbin, Richard},
+	date-added = {2025-04-01 09:01:03 +0200},
+	date-modified = {2025-04-01 09:01:03 +0200},
+	journal = {bioinformatics},
+	number = {14},
+	pages = {1754--1760},
+	publisher = {Oxford University Press},
+	title = {Fast and accurate short read alignment with Burrows--Wheeler transform},
+	volume = {25},
+	year = {2009}
+}
+
 @article{lipatov:2015,
   title={Maximum likelihood estimation of biological relatedness from low coverage sequencing data},
   author={Lipatov, Mikhail and Sanjeev, Komal and Patro, Rob and Veeramah, Krishna R},
@@ -62,6 +75,19 @@
   year={2015},
   publisher={Cold Spring Harbor Laboratory},
   doi = "10.1101/023374"
+}
+
+@article{mathieson2015genome,
+	author = {Mathieson, Iain and Lazaridis, Iosif and Rohland, Nadin and Mallick, Swapan and Patterson, Nick and Roodenberg, Song{\"u}l Alpaslan and Harney, Eadaoin and Stewardson, Kristin and Fernandes, Daniel and Novak, Mario and others},
+	date-added = {2025-04-01 09:34:18 +0200},
+	date-modified = {2025-04-01 09:34:18 +0200},
+	journal = {Nature},
+	number = {7583},
+	pages = {499--503},
+	publisher = {Nature Publishing Group UK London},
+	title = {Genome-wide patterns of selection in 230 ancient Eurasians},
+	volume = {528},
+	year = {2015}
 }
 
 @article{monroy:2018,
@@ -74,6 +100,18 @@
   year={2018},
   publisher={Public Library of Science San Francisco, CA USA}, 
   doi = "10.1371/journal.pone.0195491"
+}
+
+@article{peltzer2016eager,
+	author = {Peltzer, Alexander and J{\"a}ger, G{\"u}nter and Herbig, Alexander and Seitz, Alexander and Kniep, Christian and Krause, Johannes and Nieselt, Kay},
+	date-added = {2025-04-01 09:00:04 +0200},
+	date-modified = {2025-04-01 09:00:04 +0200},
+	journal = {Genome biology},
+	pages = {1--14},
+	publisher = {Springer},
+	title = {EAGER: efficient ancient genome reconstruction},
+	volume = {17},
+	year = {2016}
 }
 
 @article{penske:2024,
