### 2025-04-01 #7916 @jonotuke
> Contribution and authorship: Has the submitting author (@BenRohrlach) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

We have added a section on contribution to the paper.

> The example pipeline uses "a real (anonymised) ancient DNA data set" which is available in the github repository. However, I was unable to find any details regarding the dataset itself - whether it is an original dataset/published/publicly available. I would suggest inhcluding further details about the data adhering to the [JOSS Policies](https://joss.readthedocs.io/en/latest/policies.html#joss-policies) if an original dataset, and including a reference or link to the original publication otherwise - similar to how the use of a simulated genotype data has been mentioned in the paper.

Added data section to paper.
