### 2025-04-01 #7916 @jonotuke
> I suggest incorporating the example pipeline into a standalone vignette that is accessible from within RStudio.

I have moved the example pipeline to a separate vignette.
