### 2025-04-02 2ccf08b
Added new vignette on example pipeline
