### 2025-04-02 #7916 @jonotuke
> I would suggest making a separate section in the README and vignette proving guidelines for third parties looking to contribute, report issues, and seek support.

Added a section to readme.
