### 2025-04-02 995fd9c
Changed example so seperate code and output
