### 2025-04-02 #7916 @jonotuke
> Lastly, I have a few recommendations (not suggesting you must address these) regarding the README documentation. Firstly, using "path/to/eigenstat/indfile", which is a widely used format for a path holder phrase, instead of "path_to_eigenstrat_indfile" which is presently used, and lastly, separating the code and the output into two snippets so that the click-to-copy function does not copy both the code and the (commented out) output lines to the clipboard.

All done and uploaded
