### 2025-04-02 jonotuke/BREADR#3 @jonotuke
>Installation instructions: According to the reference manual (https://cran.r-project.org/web/packages/BREADR/BREADR.pdf), the package requires R version ≥ 2.10. However, I was unable to install the package from CRAN using R 4.1.2.

I have updated to a later version of R.
