### 2025-04-02 jonotuke/BREADR#3 @jonotuke
> It would be helpful to include the link to the package website (https://jonotuke.github.io/BREADR/) in the README file, so that users can more easily locate the reference manual and other documentation.

Done.
