### 2025-04-02 jonotuke/BREADR#3 @jonotuke
> Example usage: It would be helpful to include a direct link to the dataset for easier testing.

Add to example pipeline vignette.
