### 2025-04-02 jonotuke/BREADR#3 @jonotuke
> Automated tests: Automated tests are present, but currently failing. I recommend the authors fix the test suite.

Done
