### 2025-04-02 jonotuke/BREADR#4 @jonotuke
We have added a link in the paper to the article that was used for the simulation.
