### 2025-04-03 #7916 @nilabhrardas
Hi, @spholmes,

I am happy to recommend this paper for acceptance.

@jonotuke, well done on implementing the suggested changes/additions to the package documentation as well as the paper.

Wish you all the best.
@nilabhrardas
