### 2025-04-03 #7916 @jonotuke
Hi all,

Sorry to ask so late, but do you mind if we add a funding statement to the paper?

@jonotuke
