### 2025-04-03 #7916 @spholmes
@jonotuke : on the contrary it is part of JOSS recommendations:
"Acknowledgement of any financial support."
