### 2025-04-05 jonotuke/BREADR#3 @xin-huang
Unfortunately, I am still unable to install the package using R 4.2.3 and the command:

```
install.packages("BREADR")
```

I encountered the following errors:

```
ERROR: dependencies ‘cowplot’, ‘ggplot2’, ‘MASS’, ‘Matrix’ are not available for package ‘doBy’
ERROR: dependencies ‘lme4’, ‘MASS’, ‘Matrix’, ‘doBy’ are not available for package ‘pbkrtest’
ERROR: dependencies ‘MASS’, ‘mgcv’, ‘pbkrtest’, ‘quantreg’, ‘lme4’ are not available for package ‘car’
ERROR: dependency ‘car’ is not available for package ‘rstatix’
ERROR: dependencies ‘ggplot2’, ‘ggrepel’, ‘ggsci’, ‘cowplot’, ‘ggsignif’, ‘rstatix’ are not available for package ‘ggpubr’
ERROR: dependencies ‘data.table’, ‘ggplot2’, ‘ggpubr’, ‘MASS’ are not available for package ‘BREADR’
```

Some packages that appear necessary for installation, such as `Matrix`, are not explicitly declared under Imports or Depends. It would be helpful to add them to the `DESCRIPTION` file to avoid installation failures. It would also be helpful if the required R version and necessary dependencies could be mentioned in the README or on the [BREADR website](https://jonotuke.github.io/BREADR/), so that users are aware before installation.

Moreover, when I tried to install Matrix using `install.packages("Matrix")`, I received the following message:

```
package ‘Matrix’ is not available for this version of R
‘Matrix’ version 1.7-3 is in the repositories but depends on R (>= 4.4)
‘Matrix’ version 1.7-3 is in the repositories but depends on R (>= 4.6)
‘Matrix’ version 1.7-3 is in the repositories but depends on R (>= 4.5)
```

So it appears that the actual R version requirement is ≥ **4.4**, not ≥ 4.2 as currently specified in the `DESCRIPTION` file. This should be updated accordingly to prevent installation failures.

I also noticed that in the [DESCRIPTION](https://github.com/jonotuke/BREADR/blob/376d901609c967197211cbe8fd7dd8de24cd7c79/DESCRIPTION#L68) file, the required R version is set to `>= 4.2`. However, the [CRAN reference manual](https://cran.r-project.org/web/packages/BREADR/BREADR.pdf) still lists the R version requirement as `>= 2.10`. You might want to update that as well for consistency.
