### 2025-04-05 jonotuke/BREADR#3 @xin-huang
I noticed that in your GitHub Actions workflow, the package is tested with `release`, `devel`, and `oldrel-1`:

https://github.com/jonotuke/BREADR/blob/376d901609c967197211cbe8fd7dd8de24cd7c79/.github/workflows/R-CMD-check.yaml#L17-L25

Since the current release is R 4.4, this means that versions such as R 4.2.x are not actually being tested. Therefore, the declared dependency R (>= 4.2) in the DESCRIPTION file is not currently verified by the test matrix.
