### 2025-04-07 #7916 @xin-huang
Hello @spholmes

I am still unable to install the package see https://github.com/jonotuke/BREADR/issues/3.

I believe the issue should not have been closed before the reviewer was able to verify the fix.
