### 2025-04-07 #7916 @jonotuke
> I believe the issue should not have been closed before the reviewer was able to verify the fix.

@xin-huang sorry about that, I falsely assumed that I was meant to close the issue so that you were informed I had address it. I will make the changes you have suggested, and leave it open.

I have a meeting planned for tomorrow night to chat to @BenRohrlach, we will review the issues then and fix.

Sorry again
@jonotuke
