### 2025-04-08 ea228ef
Add dependencies to README, added Matrix as suggests
