+++ README.Rmd
@@ -40,6 +40,23 @@ You can install the development version of BREADR from [GitHub](https://github.c
 remotes::install_github("jonotuke/BREADR")
 ```
 
+ALso to ensure that BREADR install correctly, we suggest installing the following packages
+
+- Matrix, 
+- data.table,
+- dplyr,
+- forcats,
+- ggplot2,
+- ggpubr,
+- grDevices,
+- magrittr,
+- MASS,
+- matrixStats,
+- purrr,
+- readr,
+- stringr and 
+- tibble.
+
 ## Contributing to BREADR
 
 Please note that `BREADR` is work in progress! If you are interested in the project and want to know more please feel free to contact Jono Tuke (<simon.tuke@adelaide.edu.au>). If you find a bug or would like to see new or improved features, please open an issue on the [GitHub](https://github.com/jonotuke/BREADR) repository.
