### 2025-04-08 jonotuke/BREADR#3 @jonotuke
I have added the dependencies to the README file.

I have added the Matrix package as a Suggests. The package checking that I run will not allow Matrix to be added as an Imports as it is not use directly.

I have set the current R version requirement to >=4.4

@jonotuke
