### 2025-04-09 1008043
Changed depends to 4.3 so that GA oldrev works
