+++ README.md
@@ -36,7 +36,7 @@ You can install the development version of BREADR from
 remotes::install_github("jonotuke/BREADR")
 ```
 
-ALso to ensure that BREADR install correctly, we suggest installing the
+Also to ensure that BREADR install correctly, we suggest installing the
 following packages
 
 - Matrix,
