+++ DESCRIPTION
@@ -66,7 +66,7 @@ Imports:
     stringr,
     tibble 
 Depends: 
-    R (>= 4.4)
+    R (>= 4.3)
 LazyData: true
 Language: en-US
 VignetteBuilder: knitr
