### 2025-04-12 jonotuke/BREADR#3 @xin-huang
I still cannot install the package with R version 4.3.3, although it installs successfully with R version 4.4.3. As noted above, the `Matrix` package on CRAN now requires R ≥ 4.4. While the CI environment likely uses pre-installed or version-matched packages via RSPM or binary mirrors, this masks actual version compatibility problems. Therefore, CI passing alone is not sufficient to demonstrate practical installability—manual testing on a clean R 4.3 setup should be part of the validation process. I recommend that you explicitly state that the package requires R ≥ 4.4.

Additionally, I noticed that the [CRAN reference manual](https://cran.r-project.org/web/packages/BREADR/BREADR.pdf) still lists the R version requirement as R ≥ 2.10. This should be updated to reflect the current minimum version required for installation.
