### 2025-04-12 #7916 @xin-huang
Hello @spholmes and @jonotuke

I was able to install the package with R version 4.4.3 and successfully reproduce the example at https://jonotuke.github.io/BREADR/articles/example-pipeline.html. Therefore, I recommend accepting the paper after the author updates the DESCRIPTION file to reflect R (≥ 4.4) and ensures that the [CRAN reference manual](https://cloud.r-project.org/web/packages/BREADR/BREADR.pdf) is updated accordingly.
