+++ DESCRIPTION
@@ -66,7 +66,7 @@ Imports:
     stringr,
     tibble 
 Depends: 
-    R (>= 4.3)
+    R (>= 4.4)
 LazyData: true
 Language: en-US
 VignetteBuilder: knitr
