### 2025-04-14 87622bf
CRAN checks and fixes
