+++ NEWS.md
@@ -1,4 +1,6 @@
-# BREADR (development version)
+# BREADR 1.0.3
+
+# BREADR 1.0.4
 
 # BREADR 1.0.3
 
