+++ NEWS.md
@@ -1,3 +1,5 @@
+# BREADR (development version)
+
 # BREADR 1.0.3
 
 # BREADR 1.0.4
