+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: BREADR
 Title: Estimates Degrees of Relatedness (Up to the Second Degree) for Extreme Low-Coverage Data
-Version: 1.0.3
+Version: 1.0.3.9000
 Authors@R: 
     c(
     person(
