### 2025-04-14 #7916 @jonotuke
Hello @spholmes abd @xin-huang

DESCRIPTION updated, and CRAN updated.

Thanks

@jonotuke
