### 2025-04-21 2df676e
Added DOIs that were missing
