+++ BREADR.bib
@@ -64,7 +64,8 @@
 	publisher = {Oxford University Press},
 	title = {Fast and accurate short read alignment with Burrows--Wheeler transform},
 	volume = {25},
-	year = {2009}
+	year = {2009}, 
+  doi = "10.1093/bioinformatics/btp324"
 }
 
 @article{lipatov:2015,
@@ -111,7 +112,8 @@
 	publisher = {Springer},
 	title = {EAGER: efficient ancient genome reconstruction},
 	volume = {17},
-	year = {2016}
+	year = {2016}, 
+  doi = "10.1186/s13059-016-0918-z"
 }
 
 @article{penske:2024,
