### 2025-04-21 #7916 @jonotuke
> [@jonotuke](https://github.com/jonotuke) : Thanks for all the hard work, we are converging to the final stretch to accptance, just a few things to straighten out: in the article there seem to be some unresolved references [li2009fast] [peltzer2006eager] in the Data Description section that don't appear in the bibliography and the editorialbot request you add their suggestion of DOI as well. Best wishes, [@spholmes](https://github.com/spholmes)

@spholmes
All fixed.

Thanks
@jonotuke
