### 2025-04-21 44e2e75
Update BREADR.bib
I added a few curly brackets to protect some of the upper case letters for proper names.
