+++ BREADR.bib
@@ -33,7 +33,7 @@
 }
 
 @article{gretzinger:2024,
-  title={Evidence for dynastic succession among early Celtic elites in Central Europe},
+  title={Evidence for dynastic succession among early {Celtic elites in Central Europe}},
   author={Gretzinger, Joscha and Schmitt, Felicitas and M{\"o}tsch, Angela and Carlhoff, Selina and Lamnidis, Thiseas Christos and Huang, Yilei and Ringbauer, Harald and Knipper, Corina and Francken, Michael and Mandt, Franziska and others},
   journal={Nature Human Behaviour},
   pages={1--14},
@@ -62,7 +62,7 @@
 	number = {14},
 	pages = {1754--1760},
 	publisher = {Oxford University Press},
-	title = {Fast and accurate short read alignment with Burrows--Wheeler transform},
+	title = {Fast and accurate short read alignment with {Burrows--Wheeler} transform},
 	volume = {25},
 	year = {2009}, 
   doi = "10.1093/bioinformatics/btp324"
@@ -86,7 +86,7 @@
 	number = {7583},
 	pages = {499--503},
 	publisher = {Nature Publishing Group UK London},
-	title = {Genome-wide patterns of selection in 230 ancient Eurasians},
+  title = {Genome-wide patterns of selection in 230 ancient {Eurasians}},
 	volume = {528},
 	year = {2015}, 
   doi = "10.1038/nature16152"
@@ -118,7 +118,7 @@
 }
 
 @article{penske:2024,
-  title={Kinship practices at the early bronze age site of Leubingen in Central Germany},
+  title={Kinship practices at the early bronze age site of Leubingen in {Central Germany}},
   author={Penske, Sandra and K{\"u}{\ss}ner, Mario and Rohrlach, Adam B and Knipper, Corina and Nov{\'a}{\v{c}}ek, Jan and Childebayeva, Ainash and Krause, Johannes and Haak, Wolfgang},
   journal={Scientific Reports},
   volume={14},
@@ -141,7 +141,7 @@
   doi = "10.1186/s13059-023-02847-7"
 }
 @article{pilli:2024,
-  title={Ancient DNA challenges prevailing interpretations of the Pompeii plaster casts},
+  title={Ancient DNA challenges prevailing interpretations of the {Pompeii} plaster casts},
   author={Pilli, Elena and Vai, Stefania and Moses, Victoria C and Morelli, Stefania and Lari, Martina and Modi, Alessandra and Diroma, Maria Angela and Amoretti, Valeria and Zuchtriegel, Gabriel and Osanna, Massimo and others},
   journal={Current Biology},
   volume={34},
