### 2025-04-22 #7916 @spholmes
Hi @jonotuke, it seems there is a version 1.0.3.9000 and a version 1.0.4 do you want to clean up the R package and call it version 1.04 and then I can run the set version command?
