+++ pkgdown.yml
@@ -1,0 +1,9 @@
+pandoc: 3.1.11
+pkgdown: 2.1.1
+pkgdown_sha: ~
+articles:
+  example-pipeline: example-pipeline.html
+last_built: 2025-04-22T20:56Z
+urls:
+  reference: https://jonotuke.github.io/BREADR/reference
+  article: https://jonotuke.github.io/BREADR/articles
