### 2025-04-28 #7916 @Kevin-Mattheus-Moerman
@BenRohrlach congratulations on this JOSS publication!!!

Thanks for editing @spholmes !!

And a special thank you to the reviewers: @nilabhrardas, @xin-huang !!!!
