### 2026-04-14 #7916 @jonotuke
@openjournals/dev
We would like to make some small edits to the paper as we have notices some typos. If this possible and how do we do this?
